+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `network` field can now be configured in `snfoundry.toml`
 
+#### Changed
+
+- `--network` flag can be used with `--add-profile` flag
+
 ## [0.55.0-rc.0] - 2025-12-18
 
 ### Forge